Live

+++ src/Store.java
@@ -162,7 +162,7 @@ public class Store {
 					totalRetweets = totalRetweets + originalUserRetweetsCount.get(userID);
 					rs.updateInt(2, totalRetweets);
 					rs.updateInt(4, totalRetweets / rs.getInt(3));
-					rs.updateFloat(10, (float)Math.round((totalRetweets / rs.getInt(6))*1000) / 1000); // Use dedcimals
+					rs.updateFloat(10, (float)Math.round(((float) totalRetweets / rs.getInt(6))*1000) / 1000); // Use dedcimals
 					rs.updateRow();
 				}
